OCR_pipeline/requirements.txt
streamlit
opencv-python
easyocr
pillow
groq
numpy
pandas